+++ lang/de/rooms.php
@@ -64,18 +64,32 @@ return [
         'title' => 'Historie',
     ],
     'members' => [
-        'add_user'         => 'Benutzer hinzufügen',
-        'bulk_edit_user'   => ':numberOfSelectedUsers Mitglieder bearbeiten',
-        'bulk_remove_user' => ':numberOfSelectedUsers Mitglieder löschen',
-        'edit_user'        => 'Mitglied bearbeiten',
-        'image'            => 'Bild',
-        'modals'           => [
+        'add_user'          => 'Benutzer hinzufügen',
+        'bulk_edit_user'    => ':numberOfSelectedUsers Mitglieder bearbeiten',
+        'bulk_import_users' => 'Mehrere Benutzer hinzufügen',
+        'bulk_remove_user'  => ':numberOfSelectedUsers Mitglieder löschen',
+        'edit_user'         => 'Mitglied bearbeiten',
+        'image'             => 'Bild',
+        'modals'            => [
             'add' => [
                 'add'         => 'Hinzufügen',
                 'no_options'  => 'Keine Einträge, bitte suchen Sie nach einem Benutzer.',
                 'no_result'   => 'Oops! Für diesen Suchbegriff konnten keine Benutzer gefunden werden.',
                 'select_role' => 'Bitte wählen Sie eine Rolle aus',
                 'select_user' => 'Bitte wählen Sie einen Benutzer aus',
+            ],
+            'bulk_import' => [
+                'can_import_users'          => 'Benutzer, die hinzugefügt werden können',
+                'cannot_import_users'       => 'Benutzer, die nicht hinzugefügt werden können',
+                'copied_invalid_users'      => 'Ungültige Benutzer wurden kopiert',
+                'copy_and_close'            => 'Ungültige Benutzer kopieren und Schließen',
+                'could_not_import_users'    => 'Benutzer, die nicht hinzugefügt werden konnten',
+                'list_description'          => 'Jede Zeile steht für eine eigene E-Mail-Adresse (maximal 1000)',
+                'list_placeholder'          => "max.mustermann@domain.de\nerika.mustermann@domain.de",
+                'import_importable_button'  => 'Gültige Benutzer hinzufügen',
+                'import_importable_question'=> 'Möchten Sie fortfahren und die Benutzer hinzufügen, die hinzugefügt werden können?',
+                'imported_users'            => 'Benutzer, die erfolgreich hinzugefügt wurden',
+                'label'                     => 'E-Mail-Liste'
             ],
             'edit' => [
                 'title'      => ':firstname :lastname bearbeiten',